+++ tiger-common/src/main/java/de/gematik/test/tiger/common/config/TigerConfigurationLoader.java
@@ -8,17 +8,22 @@ import static de.gematik.test.tiger.common.config.TigerConfigurationKeyString.wr
 import com.fasterxml.jackson.core.JacksonException;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.TreeNode;
+import com.fasterxml.jackson.core.Version;
 import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.Module;
 import com.fasterxml.jackson.databind.*;
+import com.fasterxml.jackson.databind.deser.DeserializationProblemHandler;
 import com.fasterxml.jackson.databind.json.JsonMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import de.gematik.test.tiger.common.TokenSubstituteHelper;
+import java.io.IOException;
 import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
+import lombok.AllArgsConstructor;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections.IteratorUtils;
@@ -30,6 +35,7 @@ import org.yaml.snakeyaml.Yaml;
 public class TigerConfigurationLoader {
 
     private ObjectMapper objectMapper;
+    private ObjectMapper strictObjectMapper;
     private List<AbstractTigerConfigurationSource> loadedSources;
     private List<TigerTemplateSource> loadedTemplates;
 
@@ -63,6 +69,14 @@ public class TigerConfigurationLoader {
             .propertyNamingStrategy(PropertyNamingStrategies.LOWER_CASE)
             .enable(MapperFeature.ACCEPT_CASE_INSENSITIVE_ENUMS)
             .addModule(new JavaTimeModule())
+            .addModule(new AllowDelayedPrimitiveResolvementModule(this))
+            .build();
+        strictObjectMapper = JsonMapper.builder()
+            .configure(MapperFeature.ACCEPT_CASE_INSENSITIVE_PROPERTIES, true)
+            .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
+            .propertyNamingStrategy(PropertyNamingStrategies.LOWER_CASE)
+            .enable(MapperFeature.ACCEPT_CASE_INSENSITIVE_ENUMS)
+            .addModule(new JavaTimeModule())
             .build();
     }
 
@@ -86,8 +100,42 @@ public class TigerConfigurationLoader {
             .findFirst();
     }
 
-    @SneakyThrows
+    /**
+     * Instantiates a bean of the given class. The base-keys denote the point from which the keys are taken.
+     * If values can not be substituted (e.g. ${key.that.does.not.exist}) they are simply returned as a string.
+     * This behaviour is more relaxed towards input errors but might delay failures from startup to runtime.
+     * <p>
+     * If the base-keys lead to a non-defined (i.e. empty) node in the tree (no values have been read) an empty
+     * optional is returned.
+     *
+     * @param configurationBeanClass The class of the configuration bean
+     * @param baseKeys Where in the configuration tree should the values be taken from?
+     * @return An instance of configurationBeanClass filled with values taken from the configuration tree
+     * @param <T>
+     */
     public <T> Optional<T> instantiateConfigurationBean(Class<T> configurationBeanClass, String... baseKeys) {
+        return instantiateConfigurationBean(configurationBeanClass, objectMapper, baseKeys);
+    }
+
+    /**
+     * Instantiates a bean of the given class. The base-keys denote the point from which the keys are taken.
+     * If values can not be substituted (e.g. ${key.that.does.not.exist}) an empty value is returned.
+     * This behaviour follows the "fail fast, fail early" approach.
+     * <p>
+     * If the base-keys lead to a non-defined (i.e. empty) node in the tree (no values have been read) an empty
+     * optional is returned.
+     *
+     * @param configurationBeanClass The class of the configuration bean
+     * @param baseKeys Where in the configuration tree should the values be taken from?
+     * @return An instance of configurationBeanClass filled with values taken from the configuration tree
+     * @param <T>
+     */
+    public <T> Optional<T> instantiateConfigurationBeanStrict(Class<T> configurationBeanClass, String... baseKeys) {
+        return instantiateConfigurationBean(configurationBeanClass, strictObjectMapper, baseKeys);
+    }
+
+    @SneakyThrows
+    private <T> Optional<T> instantiateConfigurationBean(Class<T> configurationBeanClass, ObjectMapper objectMapper, String... baseKeys) {
         initialize();
 
         TreeNode targetTree = convertToTree();
@@ -423,5 +471,61 @@ public class TigerConfigurationLoader {
 
     public ObjectMapper getObjectMapper() {
         return objectMapper;
+    }
+
+    @AllArgsConstructor
+    private static class AllowDelayedPrimitiveResolvementModule extends Module {
+
+        private TigerConfigurationLoader tigerConfigurationLoader;
+
+        @Override
+        public String getModuleName() {
+            return "fallback provider";
+        }
+
+        @Override
+        public Version version() {
+            return Version.unknownVersion();
+        }
+
+        @Override
+        public void setupModule(SetupContext setupContext) {
+            setupContext.addDeserializationProblemHandler(new ClazzFallbackConverter(tigerConfigurationLoader));
+        }
+    }
+
+    @AllArgsConstructor
+    private static class ClazzFallbackConverter extends DeserializationProblemHandler {
+
+        private TigerConfigurationLoader tigerConfigurationLoader;
+
+        @Override
+        public Object handleWeirdStringValue(DeserializationContext ctxt, Class<?> targetType, String valueToConvert, String failureMsg) throws IOException {
+            if ((valueToConvert.contains("!{") || valueToConvert.contains("${"))
+                && (TokenSubstituteHelper.substitute(valueToConvert, tigerConfigurationLoader).equals(valueToConvert))) {
+                if (targetType.equals(Boolean.class) || targetType.equals(Integer.class) || targetType.equals(Long.class) ||
+                    targetType.equals(Character.class) || targetType.equals(Double.class) || targetType.equals(Float.class) ||
+                    targetType.equals(Byte.class) || targetType.equals(Short.class)) {
+                    return null;
+                } else if (targetType.equals(boolean.class)) {
+                    return false;
+                } else if (targetType.equals(int.class)) {
+                    return -1;
+                } else if (targetType.equals(long.class)) {
+                    return (long) -1;
+                } else if (targetType.equals(double.class)) {
+                    return -1.;
+                } else if (targetType.equals(float.class)) {
+                    return -1f;
+                } else if (targetType.equals(short.class)) {
+                    return (short) -1;
+                } else if (targetType.equals(char.class)) {
+                    return ' ';
+                } else if (targetType.equals(byte.class)) {
+                    return (byte) -1;
+                }
+            }
+            return super.handleWeirdStringValue(ctxt, targetType, valueToConvert, failureMsg);
+        }
     }
 }