+++ tiger-common/src/test/java/de/gematik/test/tiger/common/config/TigerConfigurationTest.java
@@ -688,12 +688,48 @@ public class TigerConfigurationTest {
             .isEqualTo("fooBar");
     }
 
+    @Test
+    void unresolveablePrimitives_shouldBeIgnored() {
+        TigerGlobalConfiguration.reset();
+        TigerGlobalConfiguration.readFromYaml(
+            "integer: '123${this.value.does.not.exist}'\n"
+                + "b: ${this.value.does.not.exist}\n"
+                + "c: ${this.value.does.not.exist}\n"
+                + "d: ${this.value.does.not.exist}\n"
+                + "l: ${this.value.does.not.exist}\n"
+                + "s: ${this.value.does.not.exist}\n"
+                + "by: ${this.value.does.not.exist}\n"
+                + "f: ${this.value.does.not.exist}\n"
+                + "objectInt: ${this.value.does.not.exist}\n"
+                + "nestedBean.bar: ${this.value.does.not.exist}");
+        var dummyBean = TigerGlobalConfiguration.instantiateConfigurationBean(DummyBean.class)
+            .get();
+        assertThat(dummyBean.getInteger()).isEqualTo(-1);
+        assertThat(dummyBean.isB()).isFalse();
+        assertThat(dummyBean.getC()).isEqualTo(' ');
+        assertThat(dummyBean.getD()).isEqualTo(-1.0);
+        assertThat(dummyBean.getF()).isEqualTo(-1.0f);
+        assertThat(dummyBean.getL()).isEqualTo(-1l);
+        assertThat(dummyBean.getBy()).isEqualTo((byte) -1);
+        assertThat(dummyBean.getS()).isEqualTo((short) -1);
+        assertThat(dummyBean.getObjectInt()).isNull();
+        assertThat(dummyBean.getNestedBean().getBar()).isEqualTo(-1);
+    }
+
     @Data
     @Builder
     public static class DummyBean {
 
         private String string;
         private int integer;
+        private boolean b;
+        private char c;
+        private double d;
+        private float f;
+        private long l;
+        private short s;
+        private byte by;
+        private Integer objectInt;
         private NestedBean nestedBean;
     }